ajustes

+++ src/SGP.Domain/Entities/Pais.cs
@@ -5,7 +5,7 @@ namespace SGP.Domain.Entities
 {
     public class Pais : BaseEntity
     {
-        public Pais(string nome, string sigla, int bacen)
+        public Pais(string nome, string sigla, short bacen)
         {
             Nome = nome;
             Sigla = sigla;
@@ -18,7 +18,7 @@ namespace SGP.Domain.Entities
 
         public string Nome { get; private set; }
         public string Sigla { get; private set; }
-        public int Bacen { get; private set; }
+        public short Bacen { get; private set; }
 
         public IReadOnlyList<Estado> Estados { get; private set; }
     }
